Moved startmenu into taskbar folder + updated imports + clock component
