+++ src/components/Taskbar/clock.jsx
@@ -1,0 +1,11 @@
+"use client";
+
+import { useState, useEffect} from 'react'
+
+export default function Clock({}) {
+  return (
+    <div>
+      
+    </div>
+  )
+}
